Fix chroot.c warning

+++ src/chroot.c
@@ -615,7 +615,7 @@ static void change_user(const struct RURI_CONTAINER *_Nonnull container)
 	ruri_log("{base}Supplementary groups: \n");
 	int ngroups = getgroups(0, NULL);
 	if (ngroups > 0) {
-		gid_t *groups = malloc(ngroups * sizeof(gid_t));
+		gid_t *groups = malloc((size_t)ngroups * sizeof(gid_t));
 		getgroups(ngroups, groups);
 		for (int i = 0; i < ngroups; i++) {
 			ruri_log("{base}%d \n", groups[i]);
